added the feature request and more

+++ app/controllers/posts_controller.rb
@@ -21,6 +21,10 @@ class PostsController < ApplicationController
     end
   end
 
+  def unproductive
+    #marking a post as unproductive
+  end
+
   def new
     #show a form to create a new post
   end
@@ -34,13 +38,13 @@ class PostsController < ApplicationController
     #if the user has made new post in the last week that was not a reply post, dont let them post a new post
     if post_params[:reply_to].nil? 
       #check to make sure the last post made by this user longer than 1 week ago
-      @user_posts = Post.where('user_id = ? AND reply_to = ? AND created_at BETWEEN ? AND ?', current_user.id, nil, (Date.current - 7), (Date.current))
-      if @user_posts.nil?
+      p = Post.where('created_at >= ?', Date.current.advance(days: -6))
+      week_posts = p.find_by('user_id = ? AND original = ? AND publish = ?', current_user.id, true, true)
+      if week_posts.nil?
         #create a new post
         @post = Post.new(
           title: post_params[:title],
           stand: post_params[:stand],
-          why: post_params[:why],
           user_id: current_user.id,
           publish: post_params[:publish],
           original: true
@@ -55,7 +59,23 @@ class PostsController < ApplicationController
           redirect_to request.referrer, alert: "something went wrong. please try again."
         end
       else
-        redirect_to request.referrer, alert: "you've already made a new post for this week. you can still react to posts though."
+                #create a new post and save it as unpublished
+                @post = Post.new(
+                  title: post_params[:title],
+                  stand: post_params[:stand],
+                  user_id: current_user.id,
+                  publish: false,
+                  original: true
+                )
+        if @post.save
+          if post_params[:publish] == true
+            redirect_to request.referrer, notice: "you've already made a new post for this week so this post was saved to your drafts for posting later. you can upgrade to premium to create more posts."
+          else
+            redirect_to request.referrer, notice: "your voice is saved in drafts."
+          end 
+        else
+          redirect_to request.referrer, alert: "something went wrong. please try again."
+        end
       end 
     else
       #check to make sure the user hasn't already replied to this post
@@ -65,15 +85,13 @@ class PostsController < ApplicationController
         @post = Post.new(
           reply_to: post_params[:reply_to],
           agree: post_params[:agree],
-          title: post_params[:title],
           stand: post_params[:stand],
-          why: post_params[:why],
           user_id: current_user.id,
           publish: post_params[:publish],
           original: false
         )
         if @post.save
-          if @post.publish?
+          if post_params[:publish] == true
             redirect_to request.referrer, notice: "your voice was sent to the world."
           else
             redirect_to request.referrer, notice: "your voice is saved in drafts."
@@ -91,24 +109,37 @@ class PostsController < ApplicationController
   def update
     #allow the user to change his stand on his post (agree or disagree)
     #allow the user the edit unpublished posts
+    @post = Post.find_by(id: params[:id])
+    if @post.user_id == current_user.id && !@post.publish?
+      if @post.publish == true && !Post.where('user_id = ? AND original = ? AND publish = ? AND created_at BETWEEN ? AND ?', current_user.id, true, true, (Date.current - 6), (Date.current)).nil?
+        @post.update(post_params)
+        @post.publish = false
+        @post.save
+        redirect_to request.referrer, alert: "You have already posted for this week. you can upgrade to premium if you would like to post more."
+      else
+        @post.update(post_params)
+        redirect_to request.referrer, notice: "Saved..."
+      end
+    end
   end
 
   def save
     save_check = SavedPost.find_by(user_id: current_user.id, post_id: params[:id])
-    if save_check.blank?
-      @save = SavedPost.new(
-        user_id: current_user.id,
-        post_id: params[:id]
-      )
-      if @save.save
-        redirect_to request.referrer, notice: "saved"
+    
+      if save_check.blank?
+        @save = SavedPost.new(
+          user_id: current_user.id,
+          post_id: params[:id]
+        )
+        if @save.save
+          redirect_to request.referrer, notice: "saved"
+        else
+          redirect_to request.referrer, notice: "something went wrong. try again."
+        end 
       else
-        redirect_to request.referrer, notice: "something went wrong. try again."
-      end 
-    else
-      save_check.destroy
-      redirect_to request.referrer, notice: "unsaved"
-    end
+        save_check.destroy
+        redirect_to request.referrer, notice: "unsaved"
+      end
   end
 
   def show
@@ -125,34 +156,38 @@ class PostsController < ApplicationController
     @posts= Post.where(user_id: current_user.id, publish: true).where(original: false)
     elsif params[:page] == "drafts"
     #show the drafts
-    @posts = Post.where(user_id: current_user.id, publish: nil)
+    @posts = Post.where(user_id: current_user.id, publish: false)
     else 
     #show the posts
-    @posts = Post.where(user_id: current_user.id, publish: true, reply_to: nil)
+    @posts = Post.where(user_id: current_user.id, publish: true)
     end
   end
 
   def follow 
     fol = Follower.find_by(follower_id: current_user.id, user_id: params[:id])
-    if fol.blank?
-      @follower = Follower.new(
-        user_id: params[:id],
-        follower_id: current_user.id
-      )
-      if @follower.save
-        redirect_to request.referrer, notice: "you are now following this voice."
+    if User.find_by(id: params[:id]) != current_user.id
+      if fol.blank?
+        @follower = Follower.new(
+          user_id: params[:id],
+          follower_id: current_user.id
+        )
+        if @follower.save
+          redirect_to request.referrer, notice: "you are now following this voice."
+        else
+          redirect_to request.referrer, alert: "something went wrong. please try again."
+        end
       else
-        redirect_to request.referrer, alert: "something went wrong. please try again."
+        fol.destroy
+        redirect_to request.referrer, notice: "you unfollowed this voice."
       end
     else
-      fol.destroy
-      redirect_to request.referrer, notice: "you unfollowed this voice."
-    end
+      redirect_to request.referrer, alert: "you cannot follow your self."
+    end 
   end
 
   private
 
   def post_params
-    params.require(:post).permit(:title, :stand, :why, :reply_to, :agree, :change_stand)
+    params.require(:post).permit(:title, :stand, :reply_to, :publish, :agree, :change_stand)
   end
 end
